Move Settings to settings.hpp
