+++ include/utils/config/types.hpp
@@ -11,6 +11,8 @@
 #include <variant>
 #include <vector>
 
+#include <utils/config/settings.hpp>
+
 /// \brief A Mapping that can be used to map a module or implementation to a specific EVSE or optionally to a Connector
 struct Mapping {
     int evse;                     ///< The EVSE id
@@ -93,30 +95,6 @@ enum class Datatype {
     Integer,
     Boolean,
     Path
-};
-
-struct Settings {
-    fs::path prefix;
-    fs::path config_file;
-    fs::path configs_dir;
-    fs::path schemas_dir;
-    fs::path modules_dir;
-    fs::path interfaces_dir;
-    fs::path types_dir;
-    fs::path errors_dir;
-    fs::path www_dir;
-    fs::path logging_config_file;
-    int controller_port;
-    int controller_rpc_timeout_ms;
-    std::string mqtt_broker_socket_path;
-    std::string mqtt_broker_host;
-    int mqtt_broker_port;
-    std::string mqtt_everest_prefix;
-    std::string mqtt_external_prefix;
-    std::string telemetry_prefix;
-    bool telemetry_enabled;
-    bool validate_schema;
-    std::string run_as_user;
 };
 
 struct ConfigurationParameterCharacteristics {
